+++ pkg/kubelet/dockershim/securitycontext/doc.go
@@ -15,4 +15,4 @@ limitations under the License.
 */
 
 // Package securitycontext contains security context api implementations
-package securitycontext // import "k8s.io/kubernetes/pkg/kubelet/dockertools/securitycontext"
+package securitycontext // import "k8s.io/kubernetes/pkg/kubelet/dockershim/securitycontext"
